tests/wpt/tests/css/css-text/parsing/line-break-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing line-break with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-line-break">
<meta name="assert" content="line-break supports the full grammar 'auto | loose | normal | strict | anywhere'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("line-break", "auto");
test_valid_value("line-break", "loose");
test_valid_value("line-break", "normal");
test_valid_value("line-break", "strict");
test_valid_value("line-break", "anywhere");
</script>
</body>
</html>
